more function

+++ f_print2.c
@@ -6,10 +6,10 @@
  *
  * Return: binary number.
  */
-int tobin(unsigned int arg)
+double tobin(unsigned int arg)
 {
-int  binarynum = 0;
-int  rem, temp = 1;
+double  binarynum = 0;
+double  rem, temp = 1;
 while (arg != 0)
 {
 rem = arg % 2;
